requirements.txt
# Zap – voice-only student assistant. No UI.
# Python 3.11+

python-dotenv>=1.0.0
requests>=2.28.0
pyaudio>=0.2.14
openai-whisper>=20231117
faster-whisper>=1.0.0
numpy>=1.24.0
piper-tts>=1.2.0
pvporcupine>=3.0.0
pvrecorder>=1.2.0
pygame>=2.5.0

# Google APIs (Calendar, Docs, Gmail)
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
dateparser>=1.2.0

# Ollama: run `ollama pull` for your model (see .env)
